feat: create user registration form, refactor user views

+++ user/views.py
@@ -2,12 +2,13 @@ import os
 
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.decorators import login_required
-from django.contrib.auth.models import User
+from django.db import transaction
 from django.http import HttpResponseNotFound
 from django.shortcuts import render, redirect
 import animals.models
-from user.forms import UpdateUserForm, AddUserMedia
+from user.forms import UpdateUserForm, AddUserMedia, RegistrationForm
 from user.models import UserMedia
+
 
 @login_required
 def index(request):
@@ -21,25 +22,18 @@ def index(request):
 
 def user_login(request):
     if request.user.is_authenticated:
-        return redirect('/')
+        return redirect('welcome')
+
     if request.method == 'POST':
-        username = request.POST.get('username')
-        password = request.POST.get('password')
         user = authenticate(
-            username=username,
-            password=password
+            username=request.POST.get('username'),
+            password=request.POST.get('password')
         )
         if user is not None:
             login(request, user)
-
-            next_url = request.GET.get('next')
-            if next_url:
-                return redirect(next_url)
-
-            return redirect("/")
+            return redirect(request.GET.get('next', 'welcome'))
         else:
             return HttpResponseNotFound("Failed to log in")
-
 
     return render(request, 'user/user_login.html', {})
 
@@ -47,46 +41,37 @@ def user_login(request):
 @login_required
 def user_logout(request):
     logout(request)
-
-    return redirect("/login")
+    return redirect("login")
 
 
 def user_register(request):
     if request.user.is_authenticated:
-        return redirect('/')
+        return redirect('welcome')
+
     if request.method == 'POST':
-        try:
-            username = request.POST.get('username')
-            email = request.POST.get('email')
-            password = request.POST.get('password')
-            new_user = User.objects.create_user(username, email, password)
-            new_user.save()
-            return redirect('/login')
-        except Exception:
-            error_message = "Username already exists"
+            form = RegistrationForm(request.POST)
+            if form.is_valid():
+                form.save()
+                return redirect('login')
 
-            return render(request, 'user/user_register.html', {"error_message": error_message}, status=409)
-
-    return render(request, 'user/user_register.html', {})
+    return render(request, 'user/user_register.html', {"form": RegistrationForm()})
 
 
 @login_required
 def user_history(request):
-    if not request.user.is_authenticated:
-        return redirect('/login')
     user_schedule = animals.models.Schedule.objects.filter(user=request.user)
-
     return render(request, 'user/user_history.html', {"schedules": user_schedule})
 
 
 @login_required
 def edit_profile(request):
-    form = UpdateUserForm(request.POST or None, instance=request.user)
-    if form.is_valid():
-        form.save()
-        return redirect('user')
+    if request.method == 'POST':
+        form = UpdateUserForm(request.POST)
+        if form.is_valid():
+            form.save()
+            return redirect('user')
 
-    return render(request, 'user/edit.html', {"form": form})
+    return render(request, 'user/edit.html', {"form": UpdateUserForm()})
 
 
 @login_required
@@ -101,11 +86,10 @@ def user_images(request):
 
             return redirect('user_images')
 
-    form = AddUserMedia()
     main_image = request.user.usermedia_set.filter(main=True).first()
     all_images = request.user.usermedia_set.filter(main=False).all()
 
-    return render(request, 'user/images.html', {"form": form,
+    return render(request, 'user/images.html', {"form": AddUserMedia(),
                                                 "main_image": main_image,
                                                 "user_images": all_images})
 
@@ -113,10 +97,10 @@ def user_images(request):
 @login_required
 def set_main_image(request, user_image_id):
     if request.method == 'POST':
-        old_main_image = request.user.usermedia_set.filter(main=True).first()
-        if old_main_image:
-            old_main_image.main = False
-            old_main_image.save()
+        previous_main_image = request.user.usermedia_set.filter(main=True).first()
+        if previous_main_image:
+            previous_main_image.main = False
+            previous_main_image.save()
         user_image = UserMedia.objects.get(id=user_image_id)
         user_image.main = True
         user_image.save()
@@ -124,11 +108,15 @@ def set_main_image(request, user_image_id):
     return redirect('user_images')
 
 
+@transaction.atomic
 @login_required
 def delete_user_image(request, user_image_id):
     if request.method == 'POST':
-        user_image = UserMedia.objects.get(id=user_image_id)
-        if len(user_image.media) > 0:
+        if not request.user.usermedia_set.filter(id=user_image_id).exists():
+            return redirect('user_images')
+        user_image = request.user.usermedia_set.filter(id=user_image_id).first()
+
+        if user_image.media:
             os.remove(user_image.media.path)
         user_image.delete()
     return redirect('user_images')
